Remove srcset stuff. It was overkill and nonsense.
And it might not even have worked; browsers only pick those sizes if
it can pick the right size for the image *before it has loaded CSS*.
It is a strange decision by browsers, but that is what it is.

Instead, just render a 1280 version in posts, in WebP and JPEG, and
don't give a dozen different downsized ones. Everything supports WebP
these days and they are tiny.

While we're there, use `aspect-ratio` instead of padding trick as all
modern browsers support it.