Frontend dependencies update, mostly by binning bad software.
Ditch stylelint, because its dependency tree is insane. It is not worth
the effort to keep it up to date when I could just write CSS that is not
bad.

Use a compat build of Vue. This seems to grow the Vue stuff to a size that
makes me quite unhappy, but when I start worrying about that I'm not going
to rewrite it in Vue 3; I'm going to replace it with a pile of
`document.createElement` and throw Webpack in the bin.

Remove the webpack dev server with HMR. It works, except when it does not,
and trying to work out "is this some random HMR breakage" vs "is something
actually wrong with my code" is less effort than "F5". That alone reduces
the dependency tree by about 150 packages (!!!).

Ditch Babel; it grows the dependency tree and is barely needed these days,
least of all for the tiny amount of JS I actually use.

All in that reduces the dependency tree to ~200 packages, which is still
insane, but rather better than the 600+ it was using before. Maybe I will
go an entire week after this without a Dependabot alert in my email.